+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2023-04-20
+
+### Fixed
+- types: fixed TransferOutput interface by refactoring to UploadOutput
+- types: renamed CreateTransferParameters for UploadInput
+- types: renamed UpdateTransferParameters for UpdateTransferInput
+- params: removed uploadStatus from events emitted
+- readme: added links and more accurate documentation
+
 ## [0.0.17] - 2023-04-18
 
 ### Fixed